% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_biolog_all.R
\name{bind_biolog_all}
\alias{bind_biolog_all}
\title{bind_biolog_all}
\usage{
bind_biolog_all(filename, sheets)
}
\arguments{
\item{filename}{the filename to read in (must be a .xlsx file)}

\item{sheets}{the sheets to read in}
}
\value{
a dataframe containing the name of the sheet, an id column (numbered from 1 - number of plates in sheet) and 96 columns representing the OD of in each well containing that substrate
}
\description{
binds biolog plate data across multiple sheets within a single file
}
