% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_biolog_sheet.R
\name{bind_biolog_sheet}
\alias{bind_biolog_sheet}
\title{bind_biolog_sheet}
\arguments{
\item{x}{the sheet to bind biolog data from}

\item{file}{the filename to read in (must be a .xlsx file)}
}
\value{
a dataframe containing the name of the sheet, an id column (numbered from 1 - number of plates in sheet) and 96 columns representing the OD of in each well containing that substrate
}
\description{
binds a single sheet of data from a biolog plate into a dataframe
}
