% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CFU_ml.R
\name{calc_CFU_ml}
\alias{calc_CFU_ml}
\title{calc_CFU_ml}
\usage{
calc_CFU_ml(plate_counts, dil_fac, vol_plated)
}
\arguments{
\item{plate_counts}{vector of bacterial plate counts}

\item{dil_fac}{the dilution factor plated. Can be a single number or a vector the same length as plate_counts}

\item{vol_plated}{the volume plated in microlitres}
}
\value{
a vector the same length as plate_counts
}
\description{
back calculate number of CFUs per mL from plate counts
}
\details{
calculated as \eqn{PFU/mL = plate count/(dilution factor x volume added)}
}
