% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PFU_ml.R
\name{calc_PFU_ml}
\alias{calc_PFU_ml}
\title{calc_PFU_ml}
\usage{
calc_PFU_ml(plate_counts, dil_fac, vol_phage)
}
\arguments{
\item{plate_counts}{vector of bacterial plate counts}

\item{dil_fac}{the dilution factor plated. Can be a single number or a vector the same length as plate_counts}

\item{vol_phage}{the volume of phage added in microlitres}
}
\value{
a vector the same length as plate_counts
}
\description{
back calculate number of plaque forming units (PFUs) per mL from phage assays
}
\details{
calculated as \eqn{PFU/mL = plate count/(dilution factor x volume added)}
}
