% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_raw_read_setup.R
\name{dada2_raw_read_setup}
\alias{dada2_raw_read_setup}
\title{dada2_raw_read_setup}
\usage{
dada2_raw_read_setup(
  packages = c("ggplot2", "dada2", "phyloseq", "DECIPHER", "tidyr", "dplyr",
    "MicrobioUoE"),
  raw_path = "data/raw_fastq",
  filt_path = "data/filtered_fastq",
  plot_path = "plots",
  output_path = "data/output",
  progress_path = "data/progress",
  ref_fasta = "data/ref_trainsets/rdp_train_set_16.fa",
  ref_fasta_spp = "data/ref_trainsets/rdp_species_assignment_16.fa",
  meta_data = "data/metadata.csv",
  fwd_error = NULL,
  rev_error = NULL,
  run_filter = "Y"
)
}
\arguments{
\item{packages}{list of packages to load in. Defaults to \code{dada2}, \code{MicrobioUoE}, \code{ggplot2}, \code{tidyr}, \code{dplyr}, \code{phyloseq}, \code{DECIPHER}}

\item{raw_path}{the path to where the raw fastq files are. Defaults to \code{data/raw_fastq}}

\item{filt_path}{the path to where the filtered fastq files should be put. Defaults to \code{data/filtered_fastq}}

\item{plot_path}{path to where the plots will be saved. Defaults to \code{plots}}

\item{output_path}{path to where output will be saved. Defaults to \code{data/output}}

\item{progress_path}{path to where progress files will be saved. Defaults to \code{data/progress}}

\item{ref_fasta}{path to reference fasta file for taxonomic assignment. Defaults to \code{data/ref_trainsets/rdp_train_set_16.fa}}

\item{ref_fasta_spp}{path to reference fasta file for species assignment. Defaults to \code{data/ref_trainsets/rdp_species_assignment_16.fa}}

\item{meta_data}{path to metadata file. Defaults to \code{data/metadata.csv}}

\item{fwd_error}{path to forward error rate if previously estimated. Ignores by default}

\item{rev_error}{path to reverse error rate if previously estimated. Ignores by default}

\item{run_filter}{whether or not to run the filtering step. Defaults to 'Y' for yes.}
}
\value{
a bunch of paths and creates the necessary folders to make sure the environment is prepared for running the dada2 pipeline
}
\description{
Function to setup the workspace for the AB_dada2_pipeline
}
\details{
sets up the environment at the beginning of \code{raw_read_processing.R} and \code{big_data_processing.R} so that the scripts run well. Assigns all the necessary paths and loads in necessary packages and data
}
\author{
Daniel Padfield
}
