% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pretty_boxplot.R
\name{geom_pretty_boxplot}
\alias{geom_pretty_boxplot}
\title{geom_pretty_boxplot}
\usage{
geom_pretty_boxplot(...)
}
\arguments{
\item{...}{arguments to pass to \code{geom_boxplot()}}
}
\description{
Combines geom_boxplot() and stat_summary() from ggplot2 to make prettier boxplots
}
\details{
aesthetics and arguments are generally set in original \code{ggplot(aes())} command. Gets rid of outliers because it wants you to overlay your actual points. Does not work when \code{position_jitter_dodge()} is needed.
}
