% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter_facets.R
\name{letter_facets}
\alias{letter_facets}
\title{letter_facets}
\usage{
letter_facets(string)
}
\arguments{
\item{string}{character vector of facet labels}
}
\description{
automatically labels facets with letters before the label
}
