% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveRmdOutput.R
\name{moveRmdOutput}
\alias{moveRmdOutput}
\title{moveRmdOutput}
\usage{
moveRmdOutput(Rmd_path, Output_path)
}
\arguments{
\item{Rmd_path}{path to Rmarkdown documents}

\item{Output_path}{path to where the output (knitted) documents are going to be stored}
}
\value{
moves all of the files with the following extensions from Rmd_path to the Output_path: .html, .docx, .pdf
}
\description{
Automatically moves all .html, .pdf or .docx files from one folder to another. Designed for Rmarkdown knitted documents that are regularly rendered in the same folder as the .Rmd document
}
\details{
If arguments are missing, \code{Rmd_path} defaults to \code{Rmarkdown/Rmd_files} and \code{Output_path} defaults to \code{Rmarkdown/Output_files}
}
\examples{
\dontrun{moveRmdOutput('Rmd_files', 'Output_files')}
}
