% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_facets.R
\name{number_facets}
\alias{number_facets}
\title{number_facets}
\usage{
number_facets(string)
}
\arguments{
\item{string}{character vector of facet labels}
}
\description{
automatically labels facets with number before the label
}
