% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_install_all.R
\name{package_install_all}
\alias{package_install_all}
\title{package_install_all}
\usage{
package_install_all(cran_packages = NULL, github_packages = NULL,
  bioc_packages = NULL)
}
\arguments{
\item{cran_packages}{a list of packages to install from CRAN}

\item{github_packages}{a list of packages to install from GitHub}

\item{bioc_packages}{a list of packages to install from Bioconductor}
}
\value{
if all packages are installed successfully, the function will return  \code{'Huzzah all the packages are installed'}. If not it will return the list of packages that have not been installed
}
\description{
installs all necessary packages from cran, github and bioconductor
}
\details{
if only CRAN packages are needed then you can only supply a list of those as desired. The function assumes \code{devtools} it is used to install packages from GitHub. In addition  \code{biocLite} is used to install packages from Bioconductor. \code{Bioconductor} will need to be downloaded before this will work \url{https://www.bioconductor.org/install/}
}
\examples{
cran_packages = c('dplyr')
github_packages = c('padpadpadpad/nlsLoop')
package_install_all(cran_packages, github_packages)
}
