% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qual_profs.R
\name{plot_qual_profs}
\alias{plot_qual_profs}
\title{plot_qual_profs}
\usage{
plot_qual_profs(filename, files_Fwd, files_Rev, subsample = NULL, ...)
}
\arguments{
\item{filename}{file name of pdf}

\item{files_Fwd}{list of fastq files of the forward reads}

\item{files_Rev}{list of fastq files of the reverse reads}

\item{subsample}{number of subsamples of the data to take instead of plotting every single file. Defaults to every single file}

\item{...}{extra arguments to pass onto pdf() such as height and width}
}
\value{
a pdf of quality profiles
}
\description{
Plots quality profiles of single fastq files
}
\details{
uses \code{dada2::plotQualityProfile} to plot multiple quality profiles into a single pdf
}
\examples{
\dontrun{
files_Fwd <- 'examp_R02.fastq'
files_Rev <- 'examp_R01.fastq',
plot_qual('~/Desktop/qualplots.pdf', files_Fwd, files_Rev)}
}
