% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_sol_vol.R
\name{stock_sol_vol}
\alias{stock_sol_vol}
\title{Calculates the volume of stock solution needed when making up working
solutions.}
\usage{
stock_sol_vol(stock_sol_conc, new_sol_conc, new_sol_vol)
}
\arguments{
\item{stock_sol_conc}{The concentration of your stock solution}

\item{new_sol_conc}{The desired concentration of your new solution}

\item{new_sol_vol}{The total volume of your new solution}
}
\description{
Uses the equation:
}
\details{
stock_sol_vol = (new_sol_conc * new_sol_vol) / stock_sol_conc

To work out the volume of your original solution (stock solution) you need
to make up the desired concentration of your new solution (eg. for a working
solution)
}
\examples{

StockConc <- 100
DesiredConc <- 10
DesiredVol <- 50

stock_sol_vol(StockConc, DesiredConc, DesiredVol)


}
\author{
Daniel Padfield
}
