% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_black.R
\name{theme_black}
\alias{theme_black}
\title{Theme to turn plots to have a black background}
\usage{
theme_black(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Font family.}
}
\description{
Theme to make ggplot2 graphs have a black background. Useful for presentations
}
