% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gazettadellosport.R
\name{theme_gazettadellosport}
\alias{theme_gazettadellosport}
\title{Theme to turn plots to have a light pink background}
\usage{
theme_gazettadellosport(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Font family.}
}
\description{
Theme to make ggplot2 graphs look like they are from the sports newspaper, gazetta dello sport
}
